Add support for Bosh links in postgresql and postgresql-extensible
If the user does not provide values for the postgres address, the
template will form the address from a postgres Bosh link if one is
available.

This commit assumes that there is only one postgres database link and
will configure both postgresql and postgresql_extensible against this
single database instance.

If the user does not specify both a postgresql-extensible address (or a
link provided) *and* a minimum of one query, the extensible section will
be omitted from the config.

Co-authored-by: Bradford D. Boyle <bboyle@pivotal.io>
Co-authored-by: Trevor Yacovone <tyacovone@pivotal.io>